working room change

+++ Poker/Poker/DataModel/Pot.cs
@@ -6,7 +6,7 @@ namespace Poker.DataModel
     public class Pot
     {
         public int _id { get; set; }
-        public double _money { get; set; }
+        public int _money { get; set; }
         public int _roomId { get; set; }
     }
 }
